refactor: move brc learn pages into sites/brc-analytics (#1637)
